requirements.txt
telethon
python-dotenv
requests
googletrans-py
